refactor: extract filter to a separate file (#88)
* refactor: extract resource filtering to a separate file

* apply De Morgan's law

* fix import

* small refactor

* add TODO comment
